Handle the actor changes from gnome-shell 3.34.
In gnome-shell 3.34, the use of Main.panel.statusArea.dateMenu.actor is
deprecated.

That is because Main.panel.statusArea.dateMenu is actually an actor, and
so Main.panel.statusArea.dateMenu.actor will be going away with the next
release.

That in turn means that right now this extension is spamming the
following for every single blink action:

gnome-shell[14970]: Usage of object.actor is deprecated for DateMenuButton
                    get@resource:///org/gnome/shell/ui/environment.js:242:29
                    _MessageStyleHandler/this._toggleStyle@/home/x/.local/share/gnome-shell/extensions/notifications-alert-on-user-menu@hackedbellini.gmail.com/extension.js:166:9

This is fairly suboptimal.

The correct fix is to check to see if Main.panel.statusArea.dateMenu is
a Clutter.Actor, and if so, just use it directly, and if not, use
Main.panel.statusArea.dateMenu.actor as before.

This has been tested on gnome-shell 3.34.1 on Ubuntu 19.10.