Check to see if we are already enabled.
So, when debugging the Actor issue in the previous commit, I ran into a
case where I was getting infinite recursion loops in _countUpdated.

This was happening because we had enabled, hit an error, reloaded the
extension, and enabled again...  Which promptly set originalCountUpdated
to _countUpdated, and then went into an infinite loop.

On the one hand, this should be impossible to hit outside of some
development corner cases.

On the other hand, I already had it fixed, so I might as well commit it
and include it in the pull request.