Add support for background color changes.
This implements the 'simple background blinking' mode suggested in
issue #36.
(https://github.com/bellini666/gnome-shell-notifications-alert/issues/36)

To do this, we have added 3 new settings:

'Use alert color' toggles if we want to actually use the defined color
for alerts, this defaults true.  If false, the color of the text remains
as the theme default.

'Use alert background color' does the exact same thing, but for the
background color.  This defaults false.

And 'Alert background color' defines the background color that we want
to use, assuming that 'Use alert background color' is true.

The implementation is fairly straight forward, if SETTING_USECOLOR is
true, add the style CSS for the color setting.  If
SETTING_USEBACKGROUNDCOLOR is true, add the style CSS for the
background-color setting.

If both are false, then we go through a lot of fuss and bother to set a
blank style.